+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/functions/hive/HiveDeclarativeAggregateFunction.java
@@ -21,6 +21,8 @@ package org.apache.flink.table.functions.hive;
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.table.api.TableException;
 import org.apache.flink.table.catalog.DataTypeFactory;
+import org.apache.flink.table.expressions.Expression;
+import org.apache.flink.table.expressions.UnresolvedCallExpression;
 import org.apache.flink.table.functions.DeclarativeAggregateFunction;
 import org.apache.flink.table.functions.FunctionKind;
 import org.apache.flink.table.types.DataType;
@@ -34,6 +36,10 @@ import java.util.List;
 import java.util.Optional;
 
 import static org.apache.flink.connectors.hive.HiveOptions.TABLE_EXEC_HIVE_NATIVE_AGG_FUNCTION_ENABLED;
+
+import static org.apache.flink.table.planner.expressions.ExpressionBuilder.hiveAggDecimalPlus;
+import static org.apache.flink.table.planner.expressions.ExpressionBuilder.plus;
+import static org.apache.flink.table.types.logical.LogicalTypeRoot.DECIMAL;
 
 /**
  * API for hive aggregation functions that are expressed in terms of expressions.
@@ -85,6 +91,15 @@ public abstract class HiveDeclarativeAggregateFunction extends DeclarativeAggreg
                             logicalType.getTypeRoot(),
                             TABLE_EXEC_HIVE_NATIVE_AGG_FUNCTION_ENABLED.key(),
                             functionName));
+        }
+    }
+
+    protected UnresolvedCallExpression adjustedPlus(
+            DataType dataType, Expression arg1, Expression arg2) {
+        if (dataType.getLogicalType().is(DECIMAL)) {
+            return hiveAggDecimalPlus(arg1, arg2);
+        } else {
+            return plus(arg1, arg2);
         }
     }
