+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/module/hive/HiveModule.java
@@ -89,7 +89,7 @@ public class HiveModule implements Module {
                                     "tumble_start")));
 
     static final Set<String> BUILTIN_NATIVE_AGG_FUNC =
-            Collections.unmodifiableSet(new HashSet<>(Arrays.asList("sum", "count", "min", "max")));
+            Collections.unmodifiableSet(new HashSet<>(Arrays.asList("sum", "count", "avg", "min", "max")));
 
     private final HiveFunctionDefinitionFactory factory;
     private final String hiveVersion;
@@ -153,10 +153,6 @@ public class HiveModule implements Module {
             return getBuiltInNativeAggFunction(name.toLowerCase());
         }
 
-        if (name.equalsIgnoreCase("avg")) {
-            return Optional.of(new HiveAverageAggFunction());
-        }
-
         // We override Hive's grouping function. Refer to the implementation for more details.
         if (name.equalsIgnoreCase("grouping")) {
             return Optional.of(
@@ -216,6 +212,9 @@ public class HiveModule implements Module {
             case "count":
                 // We override Hive's sum function by native implementation to supports hash-agg
                 return Optional.of(new HiveCountAggFunction());
+            case "avg":
+                // We override Hive's avg function by native implementation to supports hash-agg
+                return Optional.of(new HiveAverageAggFunction());
             case "min":
                 // We override Hive's min function by native implementation to supports hash-agg
                 return Optional.of(new HiveMinAggFunction());
